requirements.txt
streamlit>=1.36
pandas>=2.0
pdfplumber>=0.11
openpyxl>=3.1
xlrd==2.0.1
Pillow>=10.0
numpy>=1.24
